feat(story-canvas): migrate StoryMediaObject — Task 1.3
Migrate StoryMediaObject to the new model contract (spec §3.3):
- x, y, scale, rotation: CGFloat → Double
- zIndex: Int? → Int (non-optional, default 0)
- isBackground, loop: Bool? → Bool (non-optional, default false)
- startTime, duration, fadeIn, fadeOut: Float? → Double?
- Add required aspectRatio: Double (legacy decoder falls back to 1.0)
- Add anchor: CGPoint (default 0.5,0.5) with custom nested Codable
- Add intrinsicDuration: Double? and mediaURL: String?

Implement StorySlide.effectiveSlideDuration() replacing Phase 0 stub:
rounds slide duration up to next full loop boundary when a background
looping video is present (ceil(base/videoDur)*videoDur).

Fix all consumer call sites across MeeshyUI — Float/CGFloat cast boundaries
at VideoClipBar, AudioClipBar, ClipInspector, TimelineTrack, SimpleSegment,
MediaElement, SetClipPropertyCommand apply(to audio:), SplitClipCommand,
MoveClipCommand, TrimClipCommand, TimelinePanel syncTrackToModel, and
TimelineViewModel+Plan4Helpers setClipFadeIn/setClipFadeOut.

All test suites pass (0 failures across every suite); overall harness exit
reflects /tmp exhaustion from repeated large DerivedData builds, not logic.

+++ packages/MeeshySDK/Sources/MeeshyUI/Story/DraggableMediaView.swift
@@ -82,10 +82,10 @@ public struct DraggableMediaView: View {
         mediaObject.kind == .video
     }
 
-    private var currentX: CGFloat { baseX ?? mediaObject.x }
-    private var currentY: CGFloat { baseY ?? mediaObject.y }
-    private var currentScale: CGFloat { baseScale ?? mediaObject.scale }
-    private var currentRotation: CGFloat { baseRotation ?? mediaObject.rotation }
+    private var currentX: CGFloat { baseX ?? CGFloat(mediaObject.x) }
+    private var currentY: CGFloat { baseY ?? CGFloat(mediaObject.y) }
+    private var currentScale: CGFloat { baseScale ?? CGFloat(mediaObject.scale) }
+    private var currentRotation: CGFloat { baseRotation ?? CGFloat(mediaObject.rotation) }
 
     /// Resolved aspect ratio used for sizing. Falls back to `image.size` when the parent
     /// hasn't supplied a value (free for images), then to 1.0 (square) as last resort.
@@ -155,10 +155,10 @@ public struct DraggableMediaView: View {
     // MARK: - Sync base state from binding
 
     private func syncBaseFromBinding() {
-        baseX = mediaObject.x
-        baseY = mediaObject.y
-        baseScale = mediaObject.scale
-        baseRotation = mediaObject.rotation
+        baseX = CGFloat(mediaObject.x)
+        baseY = CGFloat(mediaObject.y)
+        baseScale = CGFloat(mediaObject.scale)
+        baseRotation = CGFloat(mediaObject.rotation)
     }
 
     // MARK: - Aspect-aware sizing
@@ -251,8 +251,8 @@ public struct DraggableMediaView: View {
                 let snapped = StoryAlignmentSnap.apply(to: CGPoint(x: rawX, y: rawY))
                 baseX = snapped.x
                 baseY = snapped.y
-                mediaObject.x = snapped.x
-                mediaObject.y = snapped.y
+                mediaObject.x = Double(snapped.x)
+                mediaObject.y = Double(snapped.y)
                 dragInitialized = false
                 onDragCommitted?()
                 onDragEnd()
@@ -267,7 +267,7 @@ public struct DraggableMediaView: View {
             .onEnded { value in
                 let newScale = min(4.0, max(0.3, currentScale * value))
                 baseScale = newScale
-                mediaObject.scale = newScale
+                mediaObject.scale = Double(newScale)
                 onDragEnd()
             }
     }
@@ -280,7 +280,7 @@ public struct DraggableMediaView: View {
             .onEnded { value in
                 let newRotation = currentRotation + value.degrees
                 baseRotation = newRotation
-                mediaObject.rotation = newRotation
+                mediaObject.rotation = Double(newRotation)
                 onDragEnd()
             }
     }
